Relocate Material picker components
