+++ src/datetime/index.js
@@ -1,2 +1,6 @@
 export { default as DatePicker } from './src/components/DatePicker';
 export { default as TimePicker } from './src/components/TimePicker';
+
+export { default as MaterialDatePicker } from './src/components/MaterialDatePicker';
+export { default as MaterialTimePicker } from './src/components/MaterialTimePicker';
+export { default as DateTimePicker } from './src/components/DateTimePicker';
